+++ stencil-workspace/storybook/.storybook/preview.js
@@ -1,7 +1,18 @@
+import React from 'react';
 import { defineCustomElements } from '../../loader';
 import yourTheme from './your-theme';
+import { themes } from '@storybook/theming';
+import addons from '@storybook/addons';
+import { DocsContainer } from '@storybook/addon-docs';
+import {
+  useDarkMode
+} from 'storybook-dark-mode';
 
 defineCustomElements();
+
+// get channel to listen to event emitter
+const channel = addons.getChannel();
+
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
@@ -11,8 +22,25 @@ export const parameters = {
       date: /Date$/,
     },
   },
+  darkMode: {
+    // Set the initial theme
+    current: 'light',
+    // Override the default dark theme
+    dark: { ...themes.dark, appBg: '#252a2e' },
+    stylePreview: true
+  },
   docs: {
     theme: yourTheme,
+    container: props => {
+      const isDark = useDarkMode();
+      React.useEffect(() => {
+        document.body.setAttribute('data-mwc-theme', isDark ? 'dark' : 'light');
+      }, [isDark]);
+
+      return (
+          <DocsContainer {...props} />
+      );
+    }
   },
   options: {
     storySort: {
@@ -29,3 +57,4 @@ export const parameters = {
     'storybook/docs/panel': { index: -1 },
   }
 }
+